Fix validation traversals test.
This is a little bit awkward, because ordinarily we'd want to
reduce from the right and use a cons operation to build up the
result; however, since javascript arrays are optimized for snoc
rather than cons, we fold from the left. Unfortunately, that
means that the errors appear in reverse order.