Add fmap to Iterables and Iterators.
This should eventually replace the Iterator#map defined in the Iterators
extension class; instead of eagerly evaluating the iterator and
producing an array as a result, this method now simply produces
a new iterator that lazily applies the function to the values
produced by the the underlying iterator. This subsequently revealed
a few places where Iterators#map was being used only for its
side-effect such that Iterators#reduce or Iterators#forEach (a newly
added extension method) would be more appropriate.

This also fixes a few errors related to inlining in older Haxe
compilers.